Tests/LibWeb/Text/expected/css/scrollable-overflow-recalc-sticky.txt

scrollHeight after transform change: 490
scrollTop in grown range: 300
sticky box sticks to scrollport top: true
relayouts performed: 0
